Fixed a bug in the moment tensor source term.
The moment tensor should have an integral of the Ricker wavelet time dependency.
That was implemented via GaussFirstDerivative function and splitting the linear form b into two - for point force and moment tensor separately.
The Ricker wavelet is the second derivative of the Gaussian function, that's why the integral is called the first derivative.